+++ p2p/src/collection/actor/config.rs
@@ -1,0 +1,5 @@
+pub struct Config<O: Originator, E: Endpoint> {
+    originator: O,
+    endpoint: E,
+    mailbox_size: usize,
+}
