+++ p2p/src/collection/actor/mod.rs
@@ -1,0 +1,5 @@
+mod config;
+mod ingress;
+
+pub use config::Config;
+pub use ingress::Mailbox;
